HCS-4: add goal withdraw policy and add authorize in crud

+++ app/Http/Livewire/Goals/Entries/Destroy.php
@@ -4,14 +4,19 @@ namespace App\Http\Livewire\Goals\Entries;
 
 use App\Models\GoalEntry;
 use Illuminate\Contracts\View\View;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 
 class Destroy extends Component
 {
+    use AuthorizesRequests;
+
     public ?GoalEntry $goalEntry = null;
 
     public function save(): void
     {
+
+        $this->authorize('delete', $this->goalEntry);
 
         $this->goalEntry->delete();
