+++ app/Http/Livewire/Goals/Withdrawals/Destroy.php
@@ -4,14 +4,19 @@ namespace App\Http\Livewire\Goals\Withdrawals;
 
 use App\Models\GoalWithdraw;
 use Illuminate\Contracts\View\View;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 
 class Destroy extends Component
 {
+    use AuthorizesRequests;
+
     public ?GoalWithdraw $goalWithdraw = null;
 
     public function save(): void
     {
+
+        $this->authorize('delete', $this->goalWithdraw);
 
         $this->goalWithdraw->delete();
